+++ Packs/FeedOffice365/Integrations/FeedOffice365/FeedOffice365.py
@@ -96,7 +96,8 @@ class Client(BaseClient):
                 raise ValueError(f'Could not parse returned data to Json. \n\nError massage: {err}')
         return result
 
-    def check_indicator_type(self, indicator):
+    @staticmethod
+    def check_indicator_type(indicator):
         """Checks the indicator type.
            The indicator type can be classified as one of the following values: CIDR, IPv6CIDR, IP, IPv6 or Domain.
 
@@ -106,14 +107,9 @@ class Client(BaseClient):
         Returns:
             The type of the indicator
         """
-        if re.match(ipv4cidrRegex, indicator):
-            return FeedIndicatorType.CIDR
-        elif re.match(ipv6cidrRegex, indicator):
-            return FeedIndicatorType.IPv6CIDR
-        elif re.match(ipv4Regex, indicator):
-            return FeedIndicatorType.IP
-        elif re.match(ipv6Regex, indicator):
-            return FeedIndicatorType.IPv6
+        is_ip_indicator = FeedIndicatorType.ip_to_indicator_type(indicator)
+        if is_ip_indicator:
+            return is_ip_indicator
         elif '*' in indicator:
             return FeedIndicatorType.DomainGlob
         # domain
@@ -159,7 +155,7 @@ def fetch_indicators(client: Client, indicator_type_lower: str, limit: int = -1)
             values = item.get(indicator_type_lower)
         if values:
             for value in values:
-                type_ = client.check_indicator_type(value)
+                type_ = Client.check_indicator_type(value)
                 raw_data = {
                     'value': value,
                     'type': type_,
@@ -167,14 +163,26 @@ def fetch_indicators(client: Client, indicator_type_lower: str, limit: int = -1)
                 for key, val in item.items():
                     if key not in ['ips', 'urls']:
                         raw_data.update({key: val})
+
+                indicator_mapping_fields = {
+                    "port": argToList(item.get('tcpPorts', '')),
+                    "service": item.get('serviceArea', '')
+                }
+
+                if item.get('expressRoute'):
+                    indicator_mapping_fields["office365expressroute"] = item.get('expressRoute')
+                if item.get('category'):
+                    indicator_mapping_fields["office365category"] = item.get('category')
+                if item.get('required'):
+                    indicator_mapping_fields["office365required"] = item.get('required')
+                if item.get('notes'):
+                    indicator_mapping_fields["description"] = item.get('notes')
+
                 indicators.append({
                     "value": value,
                     "type": type_,
                     "rawJSON": raw_data,
-                    "fields": {
-                        "port": argToList(item.get('tcpPorts', '')),
-                        "subfeed": item.get('serviceArea', '')
-                    }
+                    "fields": indicator_mapping_fields
                 })
 
     return indicators