+++ Packs/FeedOffice365/Integrations/FeedOffice365/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
-Added the **DomainGlob** indicator type to the feed's output, which might be applied when domains are returned.
+- Added the **DomainGlob** indicator type to the feed's output, which might be applied when domains are returned.
+- Added mapping to new indicator fields.
+
 
 ## [20.3.1] - 2020-03-04