+++ Packs/FeedOffice365/Integrations/FeedOffice365/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## [Unreleased]
+- Added the **DomainGlob** indicator type to the feed's output, which might be applied when domains are returned.
+- Added mapping to new indicator fields.
+
+
+## [20.3.1] - 2020-03-04
 
 
 ## [20.2.4] - 2020-02-25